minor fix
I still can't figure out what is making the Serial.write() on STM32 so slow.
One suspected reason is because I am using an Arduino Nano as an USB FTDI UART device.

+++ OneByOne/FrequencyMeasuring(use_IDE_SerialMonitor)/Freq_1B1AVR/Freq_1B1AVR.ino
@@ -8,7 +8,7 @@ unsigned int timeout=durationWindow+1;
 unsigned int readValue;
 unsigned int localAvg;
 unsigned int loopCount=-1;
-unsigned int timer;
+unsigned long timer;
 
 void setup(){
   Serial.begin(230400);
